error reporting, fixed tower step ranges, names replaced numbers
Added error message showing call location if printCurrenPosition gets
bad data.
Added error returns from delta transform up to move routines.
Moved a couple defines to repetier.h
Fixed comments in configuration.h and placed both leveling flags
adjacent.
Used names for axis step flag numbers (readability.
(Not including backlash flags)
Added some convinience debugging macros, normally disabled by DEGUG
flag.
Added soft endstop customized to each tower (in case they are not
symetrical)
32 bit line numbers for long file error reporting
Changed one of leveling flags to follow naming convention "FEATURE_xxx"
reordered updatederived parameter so invaxisstepspermm was avilable for
delta derived parameters.
derive maximum cartesian z coordinate from tower heights
homezaxis simplified, guarded against wonky levelling eeprom values.

+++ src/ArduinoAVR/Repetier/gcode.h
@@ -27,7 +27,7 @@ class GCode   // 52 uint8_ts per command needed
 public:
     unsigned int M;
     unsigned int G;
-    uint16_t N; // Line numbers are only checked for the first 16 bit!
+    uint32_t N; // Line numbers are needed to report error line in large files
     float X;
     float Y;
     float Z;
